requirements.txt
flask
azure-communication-callautomation
asyncio